Make into a package, split code and tests
